requirements.txt
Django==1.9.6
mysqlclient==1.3.10
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
django-crispy-forms==1.6.0
Markdown==2.6.6

